pyproject.toml
[project]
name = "fastapi-getpaid"
version = "0.1.0"
description = "FastAPI framework adapter for getpaid payment processing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Dominik Kozaczko" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "python-getpaid-core>=0.1.0",
    "fastapi>=0.115.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
sqlalchemy = [
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "ty>=0.0.17",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
]
docs = [
    "sphinx>=7.0",
    "furo",
    "myst-parser",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_getpaid"]

[tool.uv.sources]
python-getpaid-core = { path = "../getpaid-core", editable = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 80
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = [
    "N818",
    "B008",
    "B027",
]

[tool.ty]
[tool.ty.environment]
python-version = "3.12"

[dependency-groups]
dev = [
    "ty>=0.0.17",
]
